Split decoder and encoder tests
