+++ decoder_test.go
@@ -43,7 +43,7 @@ func TestDecodeFile31(t *testing.T) {
 
 	// Test password matching
 	pw := db.Content.Root.Groups[0].Groups[0].Entries[0].GetPassword()
-	if string(pw) != "Password" {
+	if pw != "Password" {
 		t.Fatalf(
 			"Failed to decode password: should be 'Password' not '%s'",
 			pw,
@@ -52,73 +52,10 @@ func TestDecodeFile31(t *testing.T) {
 
 	// Test secondary password matching
 	pw = db.Content.Root.Groups[0].Groups[0].Entries[1].GetPassword()
-	if string(pw) != "AnotherPassword" {
+	if pw != "AnotherPassword" {
 		t.Fatalf(
 			"Failed to decode password: should be 'AnotherPassword' not '%s'",
 			pw,
-		)
-	}
-
-	//
-	// Test encode system on opened database
-	//
-	f, err := os.Create("tests/kdbx3/tmp.kdbx")
-	if err != nil {
-		t.Fatalf("Failed to open file for writing: %s", err)
-	}
-
-	// Change the value of an entry element to see if the change stays after decoding
-	db.Content.Root.Groups[0].Groups[0].Entries[0].Get("URL").Value.Content = "http://github.com"
-
-	// Lock entries
-	err = db.LockProtectedEntries()
-	if err != nil {
-		t.Fatalf("Problem locking entries. %s", err)
-	}
-
-	// Test encoding
-	enc := NewEncoder(f)
-	err = enc.Encode(db)
-	if err != nil {
-		t.Fatalf("Failed to encode file: %s", err)
-	}
-
-	// Open the new database
-	tmpfile, err := os.Open("tests/kdbx3/tmp.kdbx")
-	if err != nil {
-		t.Fatalf("Failed to open keepass file: %s", err)
-	}
-	defer tmpfile.Close()
-
-	// Decode the new database
-	db = NewDatabase()
-	db.Credentials = NewPasswordCredentials("abcdefg12345678")
-	err = NewDecoder(tmpfile).Decode(db)
-	if err != nil {
-		t.Fatalf("Failed to decode temporary file: %s", err)
-	}
-
-	// Unlock entries
-	err = db.UnlockProtectedEntries()
-	if err != nil {
-		t.Fatalf("Problem unlocking entries. %s", err)
-	}
-
-	// Test password matching
-	pw = db.Content.Root.Groups[0].Groups[0].Entries[0].GetPassword()
-	if pw != "Password" {
-		t.Fatalf(
-			"Failed to decode password: should be 'Password' not '%s'",
-			pw,
-		)
-	}
-
-	// Test new value matching
-	url := db.Content.Root.Groups[0].Groups[0].Entries[0].GetContent("URL")
-	if url != "http://github.com" {
-		t.Fatalf(
-			"Failed to decode url: should be 'http://github.com' not '%s'",
-			url,
 		)
 	}
 }
@@ -161,7 +98,7 @@ func TestDecodeFile4(t *testing.T) {
 
 	// Test password matching
 	pw := db.Content.Root.Groups[0].Groups[0].Entries[0].GetPassword()
-	if string(pw) != "Password" {
+	if pw != "Password" {
 		t.Fatalf(
 			"Failed to decode password: should be 'Password' not '%s'",
 			pw,
@@ -170,73 +107,10 @@ func TestDecodeFile4(t *testing.T) {
 
 	// Test secondary password matching
 	pw = db.Content.Root.Groups[0].Groups[0].Entries[1].GetPassword()
-	if string(pw) != "AnotherPassword" {
+	if pw != "AnotherPassword" {
 		t.Fatalf(
 			"Failed to decode password: should be 'AnotherPassword' not '%s'",
 			pw,
-		)
-	}
-
-	//
-	// Test encode system on opened database
-	//
-	f, err := os.Create("tests/kdbx4/tmp.kdbx")
-	if err != nil {
-		t.Fatalf("Failed to open file for writing: %s", err)
-	}
-
-	// Change the value of an entry element to see if the change stays after decoding
-	db.Content.Root.Groups[0].Groups[0].Entries[0].Get("URL").Value.Content = "http://github.com"
-
-	// Lock entries
-	err = db.LockProtectedEntries()
-	if err != nil {
-		t.Fatalf("Problem locking entries. %s", err)
-	}
-
-	// Test encoding
-	enc := NewEncoder(f)
-	err = enc.Encode(db)
-	if err != nil {
-		t.Fatalf("Failed to encode file: %s", err)
-	}
-
-	// Open the new database
-	tmpfile, err := os.Open("tests/kdbx4/tmp.kdbx")
-	if err != nil {
-		t.Fatalf("Failed to open keepass file: %s", err)
-	}
-	defer tmpfile.Close()
-
-	// Decode the new database
-	db = NewDatabase()
-	db.Credentials = NewPasswordCredentials("abcdefg12345678")
-	err = NewDecoder(tmpfile).Decode(db)
-	if err != nil {
-		t.Fatalf("Failed to decode temporary file: %s", err)
-	}
-
-	// Unlock entries
-	err = db.UnlockProtectedEntries()
-	if err != nil {
-		t.Fatalf("Problem unlocking entries. %s", err)
-	}
-
-	// Test password matching
-	pw = db.Content.Root.Groups[0].Groups[0].Entries[0].GetPassword()
-	if pw != "Password" {
-		t.Fatalf(
-			"Failed to decode password: should be 'Password' not '%s'",
-			pw,
-		)
-	}
-
-	// Test new value matching
-	url := db.Content.Root.Groups[0].Groups[0].Entries[0].GetContent("URL")
-	if url != "http://github.com" {
-		t.Fatalf(
-			"Failed to decode url: should be 'http://github.com' not '%s'",
-			url,
 		)
 	}
 }
@@ -285,7 +159,7 @@ func TestDecodeFile31_Key(t *testing.T) {
 
 	// Test password matching
 	pw := db.Content.Root.Groups[0].Groups[0].Entries[0].GetPassword()
-	if string(pw) != "Password" {
+	if pw != "Password" {
 		t.Fatalf(
 			"Failed to decode password: should be 'Password' not '%s'",
 			pw,
@@ -294,7 +168,7 @@ func TestDecodeFile31_Key(t *testing.T) {
 
 	// Test secondary password matching
 	pw = db.Content.Root.Groups[0].Groups[0].Entries[1].GetPassword()
-	if string(pw) != "AnotherPassword" {
+	if pw != "AnotherPassword" {
 		t.Fatalf(
 			"Failed to decode password: should be 'AnotherPassword' not '%s'",
 			pw,
@@ -346,7 +220,7 @@ func TestDecodeFile4_Key(t *testing.T) {
 
 	// Test password matching
 	pw := db.Content.Root.Groups[0].Groups[0].Entries[0].GetPassword()
-	if string(pw) != "Password" {
+	if pw != "Password" {
 		t.Fatalf(
 			"Failed to decode password: should be 'Password' not '%s'",
 			pw,
@@ -355,7 +229,7 @@ func TestDecodeFile4_Key(t *testing.T) {
 
 	// Test secondary password matching
 	pw = db.Content.Root.Groups[0].Groups[0].Entries[1].GetPassword()
-	if string(pw) != "AnotherPassword" {
+	if pw != "AnotherPassword" {
 		t.Fatalf(
 			"Failed to decode password: should be 'AnotherPassword' not '%s'",
 			pw,
@@ -363,7 +237,7 @@ func TestDecodeFile4_Key(t *testing.T) {
 	}
 }
 
-// Decode database v4.0
+// Decode database v4.0 without compression
 func TestDecodeFile4_NoCompression(t *testing.T) {
 	// Open file
 	file, err := os.Open("tests/kdbx4/example-nocompression.kdbx")
@@ -401,7 +275,7 @@ func TestDecodeFile4_NoCompression(t *testing.T) {
 
 	// Test password matching
 	pw := db.Content.Root.Groups[0].Groups[0].Entries[0].GetPassword()
-	if string(pw) != "Password" {
+	if pw != "Password" {
 		t.Fatalf(
 			"Failed to decode password: should be 'Password' not '%s'",
 			pw,
@@ -410,132 +284,10 @@ func TestDecodeFile4_NoCompression(t *testing.T) {
 
 	// Test secondary password matching
 	pw = db.Content.Root.Groups[0].Groups[0].Entries[1].GetPassword()
-	if string(pw) != "AnotherPassword" {
+	if pw != "AnotherPassword" {
 		t.Fatalf(
 			"Failed to decode password: should be 'AnotherPassword' not '%s'",
 			pw,
-		)
-	}
-
-	//
-	// Test encode system on opened database
-	//
-	f, err := os.Create("tests/kdbx4/tmp.kdbx")
-	if err != nil {
-		t.Fatalf("Failed to open file for writing: %s", err)
-	}
-
-	// Change the value of an entry element to see if the change stays after decoding
-	db.Content.Root.Groups[0].Groups[0].Entries[0].Get("URL").Value.Content = "http://github.com"
-
-	// Lock entries
-	err = db.LockProtectedEntries()
-	if err != nil {
-		t.Fatalf("Problem locking entries. %s", err)
-	}
-
-	// Test encoding
-	enc := NewEncoder(f)
-	err = enc.Encode(db)
-	if err != nil {
-		t.Fatalf("Failed to encode file: %s", err)
-	}
-
-	// Open the new database
-	tmpfile, err := os.Open("tests/kdbx4/tmp.kdbx")
-	if err != nil {
-		t.Fatalf("Failed to open keepass file: %s", err)
-	}
-	defer tmpfile.Close()
-
-	// Decode the new database
-	db = NewDatabase()
-	db.Credentials = NewPasswordCredentials("abcdefg12345678")
-	err = NewDecoder(tmpfile).Decode(db)
-	if err != nil {
-		t.Fatalf("Failed to decode temporary file: %s", err)
-	}
-
-	// Unlock entries
-	err = db.UnlockProtectedEntries()
-	if err != nil {
-		t.Fatalf("Problem unlocking entries. %s", err)
-	}
-
-	// Test password matching
-	pw = db.Content.Root.Groups[0].Groups[0].Entries[0].GetPassword()
-	if pw != "Password" {
-		t.Fatalf(
-			"Failed to decode password: should be 'Password' not '%s'",
-			pw,
-		)
-	}
-
-	// Test new value matching
-	url := db.Content.Root.Groups[0].Groups[0].Entries[0].GetContent("URL")
-	if url != "http://github.com" {
-		t.Fatalf(
-			"Failed to decode url: should be 'http://github.com' not '%s'",
-			url,
-		)
-	}
-}
-
-func TestCreateNewFile(t *testing.T) {
-	//Creates a brand new kdbx file using only the library
-	newdb := NewDatabase()
-
-	if newdb.Content.Root.Groups[0].Entries[0].GetTitle() != "Sample Entry" {
-		t.Fatalf(
-			"NewRootData() seemed to not work, title should be 'Sample Entry', was %s",
-			newdb.Content.Root.Groups[0].Entries[0].GetTitle(),
-		)
-	}
-
-	newdb.Credentials = NewPasswordCredentials("password")
-
-	err := newdb.LockProtectedEntries()
-	if err != nil {
-		t.Fatalf("Problem locking entries. %s", err)
-	}
-
-	newfile, err := os.Create("tests/new.kdbx")
-	if err != nil {
-		t.Fatal(err)
-	}
-	newencoder := NewEncoder(newfile)
-	err = newencoder.Encode(newdb)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log("Please open example/new.kdbx with keepass2 to verify that it works")
-
-	// Test decode of the new file
-	tmpfile, err := os.Open("tests/new.kdbx")
-	if err != nil {
-		t.Fatalf("Failed to open new keepass file: %s", err)
-	}
-	defer tmpfile.Close()
-
-	// Decode the new database
-	newdb = NewDatabase()
-	newdb.Credentials = NewPasswordCredentials("password")
-	err = NewDecoder(tmpfile).Decode(newdb)
-	if err != nil {
-		t.Fatalf("Failed to decode new keepass file: %s", err)
-	}
-
-	// Unlock entries
-	err = newdb.UnlockProtectedEntries()
-	if err != nil {
-		t.Fatalf("Problem unlocking entries. %s", err)
-	}
-
-	// Test title matching
-	if newdb.Content.Root.Groups[0].Entries[0].GetTitle() != "Sample Entry" {
-		t.Fatalf(
-			"Decoding seemed to not work, title should be 'Sample Entry', was %s",
-			newdb.Content.Root.Groups[0].Entries[0].GetTitle(),
 		)
 	}
 }
